[REFACTOR] Text 컴포넌트 리팩토링 #10
- Text를 하나의 컴포넌트로 모으고 스타일을 props로 수령
