--- src/components/atoms/texts/BoldText.tsx
@@ -1,48 +1,0 @@
-import React from 'react';
-import classNames from 'classnames';
-
-import { FontSize, FONT_SIZE_MAPS } from '@/styles/styleConstants';
-
-export enum TextColor {
-  BLACK = 'black',
-  WHITE = 'white',
-  RED = 'red',
-  LIGHTGRAY = 'lightgray',
-  DARKGRAY = 'darkgray',
-  PRIMARY = 'primary',
-}
-
-const TEXT_COLOR_MAPS = {
-  [TextColor.BLACK]: 'text-black',
-  [TextColor.WHITE]: 'text-white',
-  [TextColor.RED]: 'text-error',
-  [TextColor.LIGHTGRAY]: 'text-gray-light',
-  [TextColor.DARKGRAY]: 'text-gray-dark',
-  [TextColor.PRIMARY]: 'text-primary',
-};
-
-export interface BoldTextPropsType
-  extends React.DetailedHTMLProps<React.HTMLAttributes<HTMLElement>, HTMLElement> {
-  label: string;
-  fontSize?: FontSize;
-  textColor?: TextColor;
-}
-
-const BoldText: React.FC<BoldTextPropsType> = ({
-  className,
-  label,
-  fontSize = FontSize.SM,
-  textColor = TextColor.BLACK,
-}: BoldTextPropsType) => {
-  const classes = classNames(
-    className,
-    'font-sans',
-    'font-bold',
-    FONT_SIZE_MAPS[fontSize],
-    TEXT_COLOR_MAPS[textColor],
-  );
-
-  return <div className={classes}>{label}</div>;
-};
-
-export default BoldText;
