+++ src/components/atoms/texts/Text.stories.tsx
@@ -1,21 +1,20 @@
 import React from 'react';
 import { Story, Meta } from '@storybook/react';
 
-import BoldText, { BoldTextPropsType, TextColor } from './BoldText';
+import Text, { TextPropsType, TextType } from './Text';
 
-import { FontSize } from '@/styles/styleConstants';
+import { FontSize, TextColor } from '@/styles/styleConstants';
 
 export default {
-  title: 'Atom/Button/BoldText',
-  component: BoldText,
+  title: 'Atom/Text/Text',
+  component: Text,
 } as Meta;
 
-const Template: Story<BoldTextPropsType> = (args) => <BoldText {...args} />;
+const Template: Story<TextPropsType> = (args) => <Text {...args} />;
 
 export const Default = Template.bind({});
 Default.args = {
   label: '일반텍스트',
-  textColor: TextColor.BLACK,
 };
 
 export const DisabledText = Template.bind({});
@@ -27,13 +26,13 @@ DisabledText.args = {
 export const ErrorText = Template.bind({});
 ErrorText.args = {
   label: '에러메시지',
-  fontSize: FontSize.XL,
+  fontSize: FontSize.SM,
   textColor: TextColor.RED,
 };
 
-export const PurpleText = Template.bind({});
-PurpleText.args = {
-  label: '프라이머리',
-  fontSize: FontSize.SM,
+export const Heading01 = Template.bind({});
+Heading01.args = {
+  label: 'heading1',
   textColor: TextColor.PRIMARY,
+  textType: TextType.H1,
 };
